paramatize eks

+++ vpc/scripts/300-get-eks-cluster.sh
@@ -126,18 +126,21 @@ if [ "$kcount" -gt "0" ]; then
         terraform fmt
         echo "validate"
         terraform validate
-        terraform refresh > /dev/null
-        echo "finish refresh"
-        rm -f t*.txt
-        tcmd=`terraform output aws_eks_cluster_${cln}_vpc_id`
-        scmd=`terraform output aws_eks_cluster_${cln}_subnet_ids | tr -d '[|]|,|"'`
-        #echo "tcmd= $tcmd"
-        #echo "scmd= $scmd"
-        ../../scripts/100-get-vpc.sh $tcmd
-        for s1 in `echo $scmd` ; do
-            #echo $s1
-            ../../scripts/105-get-subnet.sh $s1
-        done
+
+        if [ $1 != "" ]; then
+            # get other stuff
+            terraform refresh > /dev/null
+            echo "finish refresh"
+            rm -f t*.txt
+            tcmd=`terraform output aws_eks_cluster_${cln}_vpc_id`
+            ../../scripts/100-get-vpc.sh $tcmd
+            #
+            scmd=`terraform output aws_eks_cluster_${cln}_subnet_ids | tr -d '[|]|,|"'`
+            for s1 in `echo $scmd` ; do
+                #echo $s1
+                ../../scripts/105-get-subnet.sh $s1
+            done
+        fi
         
     done  # k  
 fi
